+++ CHANGELOG.md
@@ -61,10 +61,11 @@
   will now utilize the new single-pixel buffer protocol. This mainly
   reduces the memory usage, but should also be slightly faster.
 * Support for high-res mouse wheel scroll events ([#1738][1738]).
+* Styled and colored underlines ([#828][828]).
 
 [1707]: https://codeberg.org/dnkl/foot/issues/1707
 [1738]: https://codeberg.org/dnkl/foot/issues/1738
-
+[828]: https://codeberg.org/dnkl/foot/issues/828
 
 ### Changed